Introduce option to disable LCS algorithm
The LCS algorithm produces excellent diffs. Unfortunately it has a
complexity, running at n^2 for the number of items in an array - which
can lead to extremely slow computations.

```
> require 'hashdiff';require 'benchmark'
> x = (1..100).map { |i| { key: i, foo: :bar } }
> puts Benchmark.measure { HashDiff.diff(x, x) }
  0.420000   0.000000   0.420000 (  0.430212)
```

If the size of the array is 1000 then we see it get really painful

```
> x = (1..1000).map { |i| { key: i, foo: :bar } }
> puts Benchmark.measure { HashDiff.diff(x, x) }
 42.680000   0.590000  43.270000 ( 43.530287)
```

This commit introduces an option to sacrifice the quality of the diff
for a faster computational result with the `use_lcs` option, which can
be set to false to disable use of the LCS algorithm.

With `use_lcs` as false the array comparison is much simpler with a
complexity of at worst 2n for an array.

```
> x = (1..100).map { |i| { key: i, foo: :bar } }
> puts Benchmark.measure { HashDiff.diff(x, x) }
  0.010000   0.000000   0.010000 (  0.004393)
> x = (1..1000).map { |i| { key: i, foo: :bar } }
> puts Benchmark.measure { HashDiff.diff(x, x, use_lcs: false) }
  0.040000   0.000000   0.040000 (  0.041610)
```

The diffs are nice for simple cases where an item is added to the start
or end of an array:

```
> HashDiff.diff([0,1,2], [0,1], use_lcs: false)
=> [["-", "[2]", 2]]
> HashDiff.diff([1,2,3,4], [1,2,3], use_lcs: false)
=> [["-", "[3]", 4]]
```

But are significantly more verbose when items change at the start and end of
arrays:

```
> HashDiff.diff([0,2,3,4,6], [0,1,2,3,4,5,6], use_lcs: false)
=> [["~", "[1]", 2, 1], ["~", "[2]", 3, 2], ["~", "[3]", 4, 3], ["~", "[4]", 6, 4], ["+", "[5]", 5], ["+", "[6]", 6]]
> HashDiff.diff([0,2,3,4,6], [0,1,2,3,4,5,6])
=> [["+", "[1]", 1], ["+", "[5]", 5]]
```